% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cerenet_feature.R
\name{plot_cerenet_feature}
\alias{plot_cerenet_feature}
\title{Bright, high-contrast spatial map for a CereNet score}
\usage{
plot_cerenet_feature(seurat, feature, palette = "inferno")
}
\arguments{
\item{seurat}{Seurat object returned by \code{run_cerenet()}}

\item{feature}{Name of the column in \code{meta.data}
(e.g. "Immune_PD11_prop", "Hypoxia1_prop")}

\item{palette}{Viridis palette ("inferno", "plasma", "magma", ...)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A ggplot2 object
}\if{html}{\out{</div>}}
}
\description{
Bright, high-contrast spatial map for a CereNet score
}
