% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cerenet.R
\name{run_cerenet}
\alias{run_cerenet}
\title{One-liner to run CereNet on a Visium folder}
\usage{
run_cerenet(path, modules, n_iter = 5)
}
\arguments{
\item{path}{Folder with 10x Visium files (outs/).}

\item{modules}{Named list: each element is a gene vector.}

\item{n_iter}{Diffusion iterations (default 5).}
}
\description{
One-liner to run CereNet on a Visium folder
}
